+++ packages/i18nify-js/src/modules/geo/index.ts
@@ -1,0 +1,2 @@
+export { default as getFlagOfCountry } from './getFlagOfCountry';
+export { default as getFlagsForAllCountries } from './getFlagsForAllCountries';