+++ packages/i18nify-js/src/modules/phoneNumber/__tests__/getDialCodes.test.ts
@@ -1,5 +1,5 @@
-import { CountryCodeType } from '../../types/geo';
-import getDialCodes from '../getDialCodes';
+import { CountryCodeType } from '../../types';
+import { getDialCodes } from '../index';
 
 describe('phoneNumber - getDialCodes', () => {
   it('should correctly map country codes to their dial codes', () => {