+++ packages/i18nify-js/src/modules/phoneNumber/__tests__/isValidPhoneNumber.test.ts
@@ -1,4 +1,4 @@
-import isValidPhoneNumber from '../isValidPhoneNumber';
+import { isValidPhoneNumber } from '../index';
 
 describe('isValidPhoneNumber', () => {
   const validTestDataSet = [