+++ packages/i18nify-js/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @razorpay/i18nify-js
 
+## 1.8.0
+
+### Minor Changes
+
+- 33a66ca: Flag service updated to return unpkg url
+
+## 1.7.0
+
+### Minor Changes
+
+- 05e4574: Flags deployed on unpkg
+
+## 1.6.1
+
+### Patch Changes
+
+- 6980dba: fix: update country code data for phone number and currency module
+
+## 1.6.0
+
+### Minor Changes
+
+- 77e9bf7: feat[ATLAS-104]: Introducing Date & Time Module
+
+## 1.5.0
+
+### Minor Changes
+
+- 92bbe13: feat: Add new Geo module - Country Flag SVG's
+
 ## 1.4.4
 
 ### Patch Changes