+++ packages/i18nify-react/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @razorpay/i18nify-react
 
+## 4.0.6
+
+### Patch Changes
+
+- Updated dependencies [33a66ca]
+  - @razorpay/i18nify-js@1.8.0
+
+## 4.0.5
+
+### Patch Changes
+
+- 05e4574: Flags deployed on unpkg
+
+### Patch Changes
+
+- Updated dependencies [05e4574]
+  - @razorpay/i18nify-js@1.7.0
+
+## 4.0.4
+
+### Patch Changes
+
+- 77e9bf7: feat[ATLAS-104]: Introducing Date & Time Module
+- Updated dependencies [77e9bf7]
+  - @razorpay/i18nify-js@1.6.0
+
+## 4.0.3
+
+### Patch Changes
+
+- Updated dependencies [92bbe13]
+  - @razorpay/i18nify-js@1.5.0
+
 ## 4.0.2
 
 ### Patch Changes